Renamed some scripts
