+++ magicctapipe/scripts/lst1_magic/__init__.py
@@ -1,4 +1,4 @@
-from .lst1_magic_dl1_to_dl2 import dl1_stereo_to_dl2
+from .lst1_magic_dl1_stereo_to_dl2 import dl1_stereo_to_dl2
 from .lst1_magic_event_coincidence import event_coincidence
 from .lst1_magic_mc_dl0_to_dl1 import mc_dl0_to_dl1
 from .lst1_magic_stereo_reco import stereo_reconstruction
@@ -7,7 +7,7 @@ from .lst1_magic_train_rfs import (
     train_direction_regressor,
     train_event_classifier,
 )
-from .magic_data_cal_to_dl1 import magic_calib_to_dl1
+from .magic_calib_to_dl1 import magic_calib_to_dl1
 from .merge_hdf_files import merge_hdf_files
 
 __all__ = [
