+++ magicctapipe/scripts/lst1_magic/lst1_magic_dl1_stereo_to_dl2.py
