+++ magicctapipe/scripts/lst1_magic/magic_calib_to_dl1.py
